requirements.txt
-r /rapidpro/pip-freeze.txt

# configuration related dependencies
django-getenv==1.3.1
django-cache-url==1.3.1
uwsgi==2.0.14
whitenoise==4.0
flower==0.9.2
